requirements.txt
pytest>=7.0.0
pytest-html>=3.0.0
requests>=2.28.0
PyYAML>=6.0
pandas>=1.5.0
openpyxl>=3.0.0 